Use PKG_CHECK_MODULES_EXT separating type of flags
in autoconf modules, PKG_CHECK_MODULES (exported from pkg.m4) put
together CFLAGS and CPPFLAGS as well as LDFLAGS and LIBS. This can be
problematic for tools whose the order of CFLAGS/CPPFLAGS and LIBS flags
matter to generate the link invocation. In particular, the order of LIBS
and -Wl,--as-needed matters a lot for the effectiveness of unused
library removal.

In consequence, a PKG_CHECK_MODULES_EXT script using the internals of
PKG_CHECK_MODULES has been written and used to output the pkg-config
result as:
* <module>_CPPFLAGS
* <module>_CFLAGS
* <module>_LDFLAGS
* <module>_LIBS