Pass datadir in CPPFLAGS rather than in config.h
Is it bad pratice to set dir in config.h at configure time. Thus we stop using
ax_define_dir in configure.ac and instead pass the datadir through AM_CPPFLAGS
(using -D). On the way, we append the package name at that moment to form the
actual PKGDATADIR.

In addition to be good practices, this changes is needed for a windows build to
be usable after make install (through msys). This indeed allows the path
conversion to be operate when the compiler is called and store the windows full
path instead of the relatively to the root of msys installation.