Register local recursive lock for gdk
Due to bug in the locking scheme of mcpanel, resizing the panel deadlock
on windows. This locking scheme (callback and rest) should of course be
fixed. But at a temporary measure, we can override the lock functions
used in gdk_threads_enter() and gdk_threads_leave() and use recursive
lock which would solve the issue.