Fix segfault when no trigger channel are selected
This ensures that processing on trigger channel data is not performed if
no trigger channel is active. This situation of not active trigger
channel can happened when trigger channel combo is cleared or when empty
channel list is added to combo.

To make this fix work, we had also to ensure that no active item is set
when empty list is added to combo box.

Change-Id: I0b1650316b1413e7cdc695d393bd4e76219f7d6d