Fix invalid data access when updating trigger input
When defining new trigger input, the label list of the trigger
channel combo is updated. This, in turn, triggers the selected channel
callback which depends on the selected_triggers and triggers array to be
allocated. Consequently, the trigger data buffer and metadata (nch,
length...) must be updated before the combo.

Refs: MMLAB-70
Change-Id: I283fb60068e207785e3c569dce2ed91b0c1a8df7