mcp_add_events() accept relative timestamp
instead of adding event relatively to the previous sample added to the
tab, now mcp_add_events() has been modified to accept an array of
mcp_event structures holding the event type and the position. The
position is the index of sample since mcp_defined_tab_input() has been
called on the same tab id.

To handle the requirement of the implementation, most of the event code
logic has been moved in scope.c

Closes: MMLAB-72
Change-Id: I9468fb469715a9aba2d424753785ff1f7526976a