delay spectrum display
When a channel just got selected, or number of point of dft is modified,
the estimation of the spectrum is bad since not enough data has been
collected.

Hence instead of displaying inacurate spectrum, only a flat zero
spectrum is shown and the normal output will be resumed once eanough
data is gathered to estimate the spectrum of the selected channel.

Fixes: #3
Change-Id: If59a2a6db96fe9235cccfcf722005e9fd10c926e