free list of selected rows safely
Result of gtk_tree_selection_get_selected_rows() is a list of
GtkTreePath* that must be freed when no longer needed. GTK documentation
mentions to free it by doing:
  g_list_foreach (list, (GFunc) gtk_tree_path_free, NULL);
  g_list_free (list);
However, this is not safe since GFunc accept 2 arguments and and
gtk_tree_path_free() has only one. By chance, this _should_ work on most
usual platform, but this will any way generate compiler warning and this
is legitimate.

Hence to fix this, free_selected_rows_list() has been added to misc.c
and implement the same logic but using an intermediate function to
ensure all pointer casts are legit.

Change-Id: I9d209bfe2c548476e9c64d8849f5a4dc778c7128